requirements.txt
nhl-api-py==2.8.0
psycopg2-binary==2.9.9
requests==2.25.1